SITES-12983: Ensuring correct and consistent display of Action Bar and Header Menu items, including submenus and icons
-- Component decomposition
