+++ src/templates/chatgpt-demo/extension-manifest.json
@@ -3,6 +3,7 @@
   "id": "chatgpt-editor-extension-demo",
   "description": "This extensions adds ChatGPT support to rich text editor in CF Editor",
   "version": "1.0.0",
+  "isDemoExtension": true,
   "templateFolder": "chatgpt-demo",
   "templateInputs": {
     "LOG_LEVEL": "debug",
