+++ src/templates/chatgpt-demo/openai-declaration.ejs
@@ -1,0 +1,29 @@
+<%_ if (extensionManifest.isDemoExtension) { -%>
+import { Configuration, OpenAIApi } from "openai"
+
+const configuration = new Configuration({
+  organization: process.env.OPENAI_ORG,
+  apiKey: process.env.OPENAI_API_KEY,
+});
+const openai = new OpenAIApi(configuration);
+
+const checkState = (state) => {
+  if (!state.selectedText) {
+    throw new Error("Please select some text first");
+  }
+}
+
+const makeRequest = async (prompt) => {
+  try {
+    const result =  await openai.createCompletion({
+      model: "text-davinci-003",
+      prompt: prompt,
+      max_tokens: 2000
+    }, { timeout: 20000 });
+    return result.data.choices[0].text;
+  } catch (e) {
+    throw new Error("An unexpected error has occurred. Please try again");
+  }
+  return "";
+}
+<%_ } -%>
