+++ src/templates/stub-app.ejs
