+++ src/templates/stub-modal.ejs
